+++ Assets/Editor Toolbox/Editor/Utilities/PropertyUtility.cs
@@ -10,6 +10,12 @@ namespace Toolbox.Editor
 {
     public static partial class PropertyUtility
     {
+        //NOTE: last non-reflection implementation was ok but support for [SerializeReference] makes it a bit slow
+        // unfortunately UnityEditor.ScriptAttributeUtility.GetFieldInfoFromProperty is internal so we have to retrive it using reflection
+        private static readonly MethodInfo getGetFieldInfoFromPropertyMethod =
+            ReflectionUtility.GetEditorMethod("UnityEditor.ScriptAttributeUtility", "GetFieldInfoFromProperty",
+                BindingFlags.NonPublic | BindingFlags.Static);
+
         /// <summary>
         /// Indicates if the property has all changes applied and can be safely used for reflection-based features.
         /// </summary>
@@ -249,7 +255,6 @@ namespace Toolbox.Editor
             return scriptInstance.GetClass();
         }
 
-
         public static FieldInfo GetFieldInfo(this SerializedProperty property)
         {
             return GetFieldInfo(property, out _);
@@ -257,14 +262,13 @@ namespace Toolbox.Editor
 
         public static FieldInfo GetFieldInfo(this SerializedProperty property, out Type propertyType)
         {
-            return GetFieldInfoFromProperty(property, out propertyType);
+            var parameters = new object[] { property, null };
+            var result = getGetFieldInfoFromPropertyMethod.Invoke(null, parameters) as FieldInfo;
+            propertyType = parameters[1] as Type;
+            return result;
         }
 
-        public static FieldInfo GetFieldInfo(this SerializedProperty property, out Type propertyType, Object target)
-        {
-            return GetFieldInfoFromProperty(property, out propertyType, target.GetType());
-        }
-
+        [Obsolete("This method is no longer safe, use GetFieldInfo() instead.")]
         public static FieldInfo GetFieldInfoFromProperty(SerializedProperty property, out Type type)
         {
             var classType = GetScriptTypeFromProperty(property);
@@ -277,8 +281,11 @@ namespace Toolbox.Editor
             return GetFieldInfoFromProperty(property, out type, classType);
         }
 
+        [Obsolete("This method is no longer safe, use GetFieldInfo() instead.")]
         public static FieldInfo GetFieldInfoFromProperty(SerializedProperty property, out Type type, Type host)
         {
+            const BindingFlags fieldFlags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic;
+
             FieldInfo field = null;
             type = host;
 
@@ -297,7 +304,6 @@ namespace Toolbox.Editor
                     continue;
                 }
 
-                const BindingFlags fieldFlags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic;
                 FieldInfo foundField = null;
                 for (var currentType = type; foundField == null && currentType != null; currentType = currentType.BaseType)
                 {
@@ -383,7 +389,7 @@ namespace Toolbox.Editor
 
         public static T GetAttribute<T>(SerializedProperty property) where T : Attribute
         {
-            return GetAttribute<T>(property, GetFieldInfoFromProperty(property, out _));
+            return GetAttribute<T>(property, GetFieldInfo(property, out _));
         }
 
         public static T GetAttribute<T>(SerializedProperty property, FieldInfo fieldInfo) where T : Attribute
@@ -393,7 +399,7 @@ namespace Toolbox.Editor
 
         public static T[] GetAttributes<T>(SerializedProperty property) where T : Attribute
         {
-            return GetAttributes<T>(property, GetFieldInfoFromProperty(property, out _));
+            return GetAttributes<T>(property, GetFieldInfo(property, out _));
         }
 
         public static T[] GetAttributes<T>(SerializedProperty property, FieldInfo fieldInfo) where T : Attribute